FW-527. cc1200.h with the 802.15.4g set up.

+++ bsp/boards/radio.h
@@ -66,7 +66,7 @@ void     radio_setFrequency(uint8_t frequency);
 void     radio_rfOn(void);
 void     radio_rfOff(void);
 // TX
-void     radio_loadPacket(uint8_t* packet, uint8_t len);
+void     radio_loadPacket(uint8_t* packet, uint16_t len);
 void     radio_txEnable(void);
 void     radio_txNow(void);
 // RX
